Add route handler for 404 requests
Add a new catch-all route which handles 404 requests. If code-splitting
is disabled and we got to this route handler, it means no section was
matched. If code-splitting was enabled and no `state.ui.chunkName` is
set, it means that no section partial route was matched and therefore no
section JS chunk is loaded.