Lots of work to get the site popover to work
Added the ability to pass along a querystring as well as a base to the SiteSelector and let the SitePopover pass it on through.
Fixed the reader controller to spin down the full post when you start from a full post screen.
Fixed the Dialog's what-to-do-on-outside-click advice. Now you return true to prevent closing the dialog instead of preventing the event. This lets clicks on a popover item that have a URL escape and change the page. This was preventing the site selector from working on full post view.
Added a style override for the site selector to make it less tall, since share buttons are usually somewhere mid page.
Turned on auto positioning for popovers for the same reason. Unsure why this was off?
Updated Popover to defer positioning and showing the content until the content had a chance to render. Without the defer, it rendered in the wrong place, because the content did not exist yet. Unsure how this ever really worked.