server/pages: Don't cache theme details
Simplify the caching logic by removing the API data cache for theme details.

As `dateUpdated` is now part of the theme details Redux subtree, and hence of
`initialState` which is used to compute the caching key, an updated theme
should still cause a cache miss and hence a freshly rendered page. This way,
the functionality provided by the API data cache is retained.

Also, markup is now returned on first hit (when cache is empty), mostly for
better DevX.